backend/requirements.txt
# ========================
# FastAPI 核心
# ========================
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ========================
# 数据库 (MySQL + SQLAlchemy)
# ========================
sqlalchemy==2.0.36
aiomysql==0.2.0
pymysql==1.1.1
alembic==1.14.0

# ========================
# 数据验证
# ========================
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# ========================
# 认证安全
# ========================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# ========================
# Redis 缓存
# ========================
redis[hiredis]==5.2.1

# ========================
# 对象存储 (MinIO/S3)
# ========================
minio==7.2.11

# ========================
# 文档处理
# ========================
PyMuPDF==1.24.14          # PDF 解析
python-docx==1.1.2        # DOCX 解析
openpyxl==3.1.5           # XLSX 解析
chardet==5.2.0             # 编码检测
Pillow==11.0.0             # 图像处理

# ========================
# LLM 客户端
# ========================
openai==1.57.2             # 兼容 OpenAI/DeepSeek/等

# ========================
# 任务队列
# ========================
celery[redis]==5.4.0
flower==2.0.1              # Celery 监控 UI

# ========================
# 工具库
# ========================
httpx==0.28.1
python-dateutil==2.9.0
orjson==3.10.12
